Changed Movement and started Note Rework
changed the movement style of the game to work off set positions and change between them using an array of empty game objects. also started the note rework by starting a new note accuracy script, adding the accuracy colliders to the player prefab and reworking the collider on the note prefab.